wip: refactor Menu components
